// src/utils/types.ts
export type SiteContent = {
  header: {
    services: string;
    projects: string;
    about: string;
    contacts: string;
  };
  hero: {
    webAgency: string;
    clearSight: string;
    brightFuture: string;
  };
  heroImage: {
    fromIdea: string;
    pixelByPixel: string;
  };
  services: {
    ourServices: string;
    tabs: {
      design: string;
      dev: string;
    };
    design: {
      header: string;
      services: {
        webDesign: {
          title: string;
          description: string;
        };
        logo: {
          title: string;
          description: string;
        };
        product: {
          title: string;
          description: string;
        };
      };
    };
    dev: {
      header: string;
      services: {
        websites: {
          title: string;
          description: string;
        };
        applications: {
          title: string;
          description: string;
        };
        eCommerce: {
          title: string;
          description: string;
        };
        business: {
          title: string;
          description: string;
        };
      };
    };
  };
  ourWork: {
    check: string;
    workData: {
      libira: {
        blockDescription: string;
        description: string;
      };
      beeTrade: {
        blockDescription: string;
        description: string;
      };
      radarDigitaly: {
        blockDescription: string;
        description: string;
      };
    };
  };
  completeSolution: {
    description: {
      from: string;
      completeSolution: string;
      ourPackage: string;
      getDetails: string;
    };
    steps: {
      wireframe: {
        description: string;
      };
      design: {
        description: string;
      };
      development: {
        description: string;
      };
      optimisation: {
        description: string;
      };
    };
  };
  footer: {
    getResponse: string;
    emailMessage: string;
    clearSight: string;
    email: string;
    message: string;
  };
};